requirements.txt
torch
numpy
matplotlib
pandas
einops
pillow
vizdoom
gymnasium>=0.29
portpicker
onnx
onnx2pytorch